Prototype/requirements.txt
streamlit
langchain
langchain-mistralai
langgraph
pypdf
python-dotenv
pytesseract
Pillow

